+++ src/main/kotlin/mine/sweeper/domain/Fields.kt
@@ -1,0 +1,8 @@
+package mine.sweeper.domain
+
+import mine.sweeper.domain.value.Field
+
+class Fields(val fieldList: List<Field>) {
+
+    val size = 25
+}
