+++ src/main/kotlin/mine/sweeper/domain/Game.kt
@@ -1,0 +1,3 @@
+package mine.sweeper.domain
+
+class Game(val fields: Fields)
