+++ src/main/kotlin/mine/sweeper/domain/Position.kt
@@ -1,0 +1,3 @@
+package mine.sweeper.domain
+
+data class Position(val x: Int, val y: Int)
