+++ src/main/kotlin/mine/sweeper/domain/value/Field.kt
@@ -1,0 +1,5 @@
+package mine.sweeper.domain.value
+
+import mine.sweeper.domain.Position
+
+class Field(val position: Position, val value: String)
