+++ src/main/kotlin/mine/sweeper/domain/value/Height.kt
@@ -1,0 +1,8 @@
+package mine.sweeper.domain.value
+
+@JvmInline
+value class Height(val value: Int) {
+    init {
+        require(value > 0)
+    }
+}
