+++ src/main/kotlin/mine/sweeper/domain/value/MineCount.kt
@@ -1,0 +1,8 @@
+package mine.sweeper.domain.value
+
+@JvmInline
+value class MineCount(val value: Int) {
+    init {
+        require(value > 0)
+    }
+}
