+++ src/main/kotlin/mine/sweeper/domain/value/Width.kt
@@ -1,0 +1,8 @@
+package mine.sweeper.domain.value
+
+@JvmInline
+value class Width(val value: Int) {
+    init {
+        require(value > 0)
+    }
+}
